+++ tests/Rule/LatteTemplatesRule/PresenterWithoutModule/Fixtures/FormsPresenter.php
@@ -6,6 +6,7 @@ namespace Efabrica\PHPStanLatte\Tests\Rule\LatteTemplatesRule\PresenterWithoutMo
 
 use Efabrica\PHPStanLatte\Tests\Rule\LatteTemplatesRule\PresenterWithoutModule\Source\CustomForm;
 use Nette\Application\UI\Form;
+use Nette\Forms\Controls\SubmitButton;
 
 final class FormsPresenter extends ParentPresenter
 {
@@ -17,7 +18,8 @@ final class FormsPresenter extends ParentPresenter
             ->setRequired();
         $form->addTextArea('textarea', 'Textarea')
             ->setHtmlAttribute('class', 'textarea-class');
-        $form->addSubmit('submit');
+        $submit = new SubmitButton();
+        $form->addComponent($submit, 'submit');
 
         $form->onSuccess[] = function (Form $form, array $values): void {
         };
@@ -36,7 +38,8 @@ final class FormsPresenter extends ParentPresenter
             ->addRule(Form::EMAIL);
         $form->addPassword('password', 'Passowrd')
             ->setRequired();
-        $form->addSubmit('submit');
+        $submit = new SubmitButton();
+        $form->addComponent($submit, 'submit');
 
         $form->onSuccess[] = function (Form $form, array $values): void {
         };
@@ -51,11 +54,67 @@ final class FormsPresenter extends ParentPresenter
         $form->addCustomText('custom_text', 'Custom text')
             ->setRequired();
         $form->addTextArea('custom_textarea', 'Custom textarea');
-        $form->addSubmit('submit');
+        $submit = new SubmitButton();
+        $form->addComponent($submit, 'submit');
 
         $form->onSuccess[] = function (Form $form, array $values): void {
         };
 
         return $form;
+    }
+
+    protected function createComponentIndirectForm(): Form
+    {
+        return $this->createIndirectForm();
+    }
+
+    protected function createComponentIndirectThroughtParentForm(): Form
+    {
+        return $this->parentCreateIndirectForm();
+    }
+
+    protected function createIndirectForm(): Form
+    {
+        $form = new Form();
+        $form->setTranslator(null);
+        $form->addSelect('select', 'Select', ['item1', 'item2', 'item3']);
+        $form->addCheckbox('checkbox', 'Checkbox');
+        $form->addText('username', 'Username')
+            ->setRequired()
+            ->addRule(Form::EMAIL);
+        $form->addPassword('password', 'Passowrd')
+            ->setRequired();
+        $submit = new SubmitButton();
+        $form->addComponent($submit, 'submit');
+
+        $form->onSuccess[] = function (Form $form, array $values): void {
+        };
+
+        return $form;
+    }
+
+    protected function createComponentIndirectFormFields(): Form
+    {
+        $form = new Form();
+        $form->setTranslator(null);
+        $this->createIndirectFormFields($form);
+
+        $form->onSuccess[] = function (Form $form, array $values): void {
+        };
+
+        return $form;
+    }
+
+    protected function createIndirectFormFields(Form $form): void
+    {
+        $form->addSelect('select', 'Select', ['item1', 'item2', 'item3']);
+        $form->addCheckbox('checkbox', 'Checkbox');
+        $form->addText('username', 'Username')
+            ->setRequired()
+            ->addRule(Form::EMAIL);
+        $form->addPassword('password', 'Passowrd')
+            ->setRequired();
+        $submit = new SubmitButton();
+        $form->addComponent($submit, 'submit');
     }
 }